+++ foodb_objectbox_adapter/test/foodb_objectbox_adapter_method_test.dart
@@ -4,7 +4,7 @@ import 'foodb_objectbox_adapter_test.dart';
 
 void main() {
   final objectBox = ObjectBoxTestContext();
-  final couchdb = CouchdbTestContext();
+  // final couchdb = CouchdbTestContext();
   // replicateBenchmarkTest(
   //     source: couchdb, target: objectBox, batchSize: 300, thousandDoc: 5);
   // group('couchdb > objectBox', () {