+++ foodb_objectbox_adapter/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.11.0 (BREAKING)
+* (breaking) update objectbox flutter dependency to 3.7
+
 ## 0.10.0
 * (fix) clear view table will remove all related view record